Some refactoring, also added more GUI-functionality

+++ src/main/java/limmen/hw2/marketplace/Item.java
@@ -16,5 +16,5 @@ public interface Item extends Remote {
     
     public String getName() throws RemoteException;
     public String getDescription() throws RemoteException;
-    public int getPrice() throws RemoteException;
+    public float getPrice() throws RemoteException;
 }
